// ui/src/openapi/models/UIFamilyMember.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ShpanKids API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ApiFamilyRole } from './ApiFamilyRole';
import {
    ApiFamilyRoleFromJSON,
    ApiFamilyRoleFromJSONTyped,
    ApiFamilyRoleToJSON,
} from './ApiFamilyRole';

/**
 * Family member
 * @export
 * @interface UIFamilyMember
 */
export interface UIFamilyMember {
    /**
     * 
     * @type {string}
     * @memberof UIFamilyMember
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UIFamilyMember
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof UIFamilyMember
     */
    lastName: string;
    /**
     * 
     * @type {ApiFamilyRole}
     * @memberof UIFamilyMember
     */
    role: ApiFamilyRole;
}

/**
 * Check if a given object implements the UIFamilyMember interface.
 */
export function instanceOfUIFamilyMember(value: object): boolean {
    if (!('email' in value)) return false;
    if (!('firstName' in value)) return false;
    if (!('lastName' in value)) return false;
    if (!('role' in value)) return false;
    return true;
}

export function UIFamilyMemberFromJSON(json: any): UIFamilyMember {
    return UIFamilyMemberFromJSONTyped(json, false);
}

export function UIFamilyMemberFromJSONTyped(json: any, ignoreDiscriminator: boolean): UIFamilyMember {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'role': ApiFamilyRoleFromJSON(json['role']),
    };
}

export function UIFamilyMemberToJSON(value?: UIFamilyMember | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'email': value['email'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'role': ApiFamilyRoleToJSON(value['role']),
    };
}